added: get all users api, updated: folder structure
